Pointer Data Reference for PI AF
This custom data reference works like a pointer. Leveraging the old
"strinc concatenation" custom data reference, this implementation allows
to dynamically point to other attributes by resolving the attribute
source path. For example, you may want Attribute1 to point to an
Attribute2 located in a sibling element which name will be returned by
Attribute3. In this case, you configure the Atrribute1 with the Pointer
DR, using this configig string:

"..\";Attribute3;"|Attribute2"

Behind the scenes,it resolves the attribute path and return the value
from that attribute.

Very useful when you can not use formulas or String Builder to point to
attributes where the path is dynamic.